app/requirements.txt
APScheduler==3.7.0
boto3==1.17.78
botocore==1.20.78
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
comtypes==1.1.9
decorator==4.4.2
Flask==2.0.1
Flask-Cors==3.0.10
idna==2.8
imageio==2.9.0
imageio-ffmpeg==0.4.3
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
MarkupSafe==2.0.1
moviepy==1.0.3
natsort==7.1.1
nltk==3.6.1
pep517==0.10.0
Pillow==8.2.0
proglog==0.1.9
pybase64==1.0.1
PyJWT==1.7.1
python-dateutil==2.8.1
pyttsx3==2.90
pytz==2019.3
PyYAML==5.4.1
regex==2021.4.4
requests==2.22.0
s3transfer==0.4.2
schedule==1.1.0
toml==0.10.2
tqdm==4.60.0
twilio==6.35.3
tzlocal==2.1
urllib3==1.25.8
Werkzeug==2.0.1
